a2a_azure/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a2a_azure"
version = "0.1.0"
authors = [
    {name = "dhaan-ish", email = "dhaanishahamed7@gmail.com"},
]
description = "A Python package for building and orchestrating AI agents with Azure OpenAI and MCP integration"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "agents", "azure", "openai", "mcp", "orchestration", "llm"]
dependencies = [
    "semantic-kernel>=1.0.0",
    "a2a>=0.1.0",
    "httpx>=0.24.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/dhaan-ish/a2a_azure"
"Bug Tracker" = "https://github.com/dhaan-ish/a2a_azure/issues"
"Documentation" = "https://github.com/dhaan-ish/a2a_azure#readme"
"Source Code" = "https://github.com/dhaan-ish/a2a_azure"

[tool.hatch.build.targets.sdist]
include = [
    "/a2a_azure",
    "/README.md",
    "/LICENSE",
]
exclude = [
    "*.pyc",
    "__pycache__",
    ".git",
    ".github",
    ".venv",
    "build",
    "dist",
    "*.egg-info",
]

[tool.hatch.build.targets.wheel]
packages = ["a2a_azure"] 

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true 